+++ mayan/apps/dynamic_search/icons.py
@@ -1,3 +1,6 @@
 from mayan.apps.appearance.classes import Icon
 
+icon_search = Icon(driver_name='fontawesome', symbol='search')
+icon_search_advanced = Icon(driver_name='fontawesome', symbol='search-plus')
+icon_search_again = Icon(driver_name='fontawesome', symbol='sync')
 icon_search_submit = Icon(driver_name='fontawesome', symbol='search')